Doc: domain hierarchy (#115)
# Background
Domains are not created equal. Some domains manage other domains. But
the problem is that every domain is handled under the same directory
`/src/domains` - which is hard to understand who is controlling what. We
now have a goal to clarify the hierarchy of domains in the API server.
After seeing the hierarchy, you can understand the relationship between
domains and how they are managed. Also you will see the naming
convention for each domain.

## What's done?
Created the following hierarchy diagram:

![image](https://github.com/user-attachments/assets/cd37324b-0135-46a8-a06c-2deb3ca29d9a)

## What's next?

### Fix inconsistent suffix for parent domain
I've learned them some of the parent domains have two suffixes:
- Chunk
- Group
And wanna fix this inconsistency later in the future. For now the
`group` sounds better as the word `chunk` is used in FE & less direct.

**Issue created: https://github.com/ajktown/api/issues/162**

### Rebrand each domain under `src/domains/ritual`
Also learned that the following hierarchy naming convention is a bit
off:

![image](https://github.com/user-attachments/assets/4f1988f4-838a-4fb2-9b5a-2fd617bf4241)

**Issue created: https://github.com/ajktown/api/issues/163**


## What are the related PRs?
- N/A

## Checklist Before PR Review
- [x] The following has been handled:
  - `Draft` is set for this PR
  - `Title` is checked
  - `Background` is filled
  - `Assignee` is set
  - `Labels` are set
  - `Development` is linked if related issue exists

## Checklist (Right Before PR Review Request)
- [x] The following has been handled:
  - Final Operation Check is done
  - Mobile View Operation Check is done
  - Make this PR as an open PR
  - `What's done` is filled & matches to the actual changes

## Checklist (Reviewers)
- [x] Review if the following has been handled:
  - Review Code
  - `Checklist Before PR Review` is correctly handled
  - `Checklist (Right Before PR Review Request)` is correctly handled
- Check if there are any other missing TODOs (Checklists) that are not
yet listed
  - Check if issue under `Development` is fully handled if exists